requirements.txt
absl-py==1.4.0
aiofiles==23.2.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
apache-beam==2.46.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
array-record==0.4.1
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
avro-python3==1.10.2
Babel==2.15.0
backports.tarfile==1.1.1
beautifulsoup4==4.12.3
bleach==6.1.0
bokeh==3.4.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
comm==0.2.2
contextlib2==21.6.0
contourpy==1.2.1
crcmod==1.7
cycler==0.12.1
Cython==3.0.10
datakit @ file:///D:/Machine%20Learning/Object_Detection_webui/Datakit
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.1.1
dm-tree==0.1.8
dnspython==2.6.1
docopt==0.6.2
docutils==0.21.2
email_validator==2.1.1
etils==1.5.2
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.3
fastavro==1.9.4
fasteners==0.19
fastjsonschema==2.19.1
ffmpy==0.3.2
filelock==3.14.0
flatbuffers==24.3.25
fonttools==4.51.0
fqdn==1.5.1
fsspec==2024.3.1
gast==0.4.0
gin-config==0.5.0
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.59.0
gradio==4.31.0
gradio_client==0.16.2
grpcio==1.63.0
h11==0.14.0
h5py==3.11.0
hdfs==2.7.3
httpcore==1.0.5
httplib2==0.21.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.0
idna==3.7
immutabledict==4.2.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
ipykernel==6.29.4
ipython==8.18.1
ipywidgets==8.1.2
isoduration==20.11.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jax==0.4.28
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.2.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.1
jupyterlab_widgets==3.0.10
kaggle==1.6.14
keras==2.10.0
Keras-Preprocessing==1.1.2
keyring==25.2.1
kiwisolver==1.4.5
libclang==18.1.1
lvis==0.5.3
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.2.0
more-itertools==10.2.0
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
nh3==0.2.17
notebook==7.2.0
notebook_shim==0.2.4
numpy==1.23.5
oauth2client==4.1.3
oauthlib==3.2.2
object_detection @ file:///D:/Machine%20Learning/Object_Detection_webui
objsize==0.6.1
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.3
overrides==7.7.0
packaging==24.0
pandas==2.2.2
pandas-bokeh==0.5.5
pandocfilters==1.5.1
parso==0.8.4
pillow==10.3.0
pkginfo==1.10.0
platformdirs==4.2.2
portalocker==2.8.2
prometheus_client==0.20.0
promise==2.3
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==3.19.0
psutil==5.9.8
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==9.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybind11==2.12.0
pycocotools==2.0.7
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pydot==1.4.2
pydub==0.25.1
Pygments==2.18.0
pymongo==3.13.0
pyparsing==2.4.7
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
python-slugify==8.0.4
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
pywinpty==2.0.13
PyYAML==5.4.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
readme_renderer==43.0
referencing==0.35.1
regex==2024.5.15
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
ruff==0.4.4
sacrebleu==2.2.0
scikit-learn==1.5.0
scipy==1.13.0
semantic-version==2.10.0
Send2Trash==1.8.3
sentencepiece==0.2.0
seqeval==1.2.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sounddevice==0.4.6
soupsieve==2.5
stack-data==0.6.3
starlette==0.37.2
tabulate==0.9.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-addons==0.20.0
tensorflow-datasets==4.9.0
tensorflow-estimator==2.10.0
tensorflow-hub==0.16.1
tensorflow-intel==2.12.0
tensorflow-io==0.31.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-metadata==1.13.0
tensorflow-model-optimization==0.8.0
tensorflow-text==2.10.0
tensorflow_object_detection_api==0.1.1
termcolor==2.4.0
terminado==0.18.1
text-unidecode==1.3
tf-keras==2.15.0
tf-models-official==2.10.1
tf-slim==1.1.0
tflite==2.10.0
tflite-support==0.4.3
threadpoolctl==3.5.0
tinycss2==1.3.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
twine==5.1.0
typeguard==2.13.3
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.9.0
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
watchfiles==0.21.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.0.3
widgetsnbextension==4.0.10
wrapt==1.14.1
xyzservices==2024.4.0
zipp==3.18.1
zstandard==0.22.0
